refactor(cli): move edit-config to top-level kairos sub-tool
edit-config touches the persistent cloud-init file, which has nothing
to do with the p2p provider that the provider binary owns. Landing
it in the provider made it reachable only through kairos provider
edit-config, which requires a provider installed and forces users of
core images to reach for a command that has no relation to what a
provider does.

Move it out of provider/internal/cli and register it as its own
top-level sub-tool of the multi-call kairos binary, so users run
it as kairos edit-config directly. The command lives in a new
editconfig/pkg/cmd package rather than under agent/, because its
concern is orthogonal to the agent's install / upgrade lifecycle
too, and future config-touching commands can join it there
without picking a sub-tool to hide under.

Refs kairos-io/kairos#4397.
