+++ editconfig/pkg/cmd/suite_test.go
@@ -1,0 +1,13 @@
+package cmd
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestEditConfigCmdSuite(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "edit-config cmd Suite")
+}
